+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.4.9
+
+* `Versions`:
+    * `Ktor`: `1.4.2` -> `1.4.3`
+* `Repo`
+    * `Common`
+        * In repos different usages of `BroadcastChannel`s has been replaced with `MutableSharedFlow`
+    * `Exposed`
+        * `asObject` open fun has been added in CRUD realization
+
 ## 0.4.8
 
 * `Versions`: